Moved require to config tab, moved file tabs to top
