+++ static/css/style.css
@@ -77,7 +77,6 @@ body {
     float:left;
     min-width: 260px;
     padding-top: 4px;
-    overflow-y: auto;
 }
 #panelfiles_gui {
     height: 31px !important;
@@ -298,6 +297,7 @@ body {
 #config {
     width: 100%;
     height: 100%;
+    overflow-y: auto;
 }
 .config {
     padding: 8px 6px;
@@ -332,10 +332,13 @@ body {
 .config .delete{
     font-size: 14px;
     float:right;
+    margin-right: 24px;
+    padding: 3px;
+    border-radius: 2px;
 }
 .config .delete:hover {
     font-size: 12px;
-    color: #f00;
+    background-color: #f00;
 }
