+++ templates/base.html
@@ -93,11 +93,14 @@
         <div id="files"></div>        
     </div>
     <div id="panel">
-        <div id="require">
-            
-        </div>
+
         <div id="edit"></div>
-        <div id="config"></div>
+        <div id="config">
+            
+            <div id="require">
+
+            </div>
+        </div>
     </div>
     
     <div id="panel_gui">
